build: bump to C++26 + small surface refactors
Apple Clang 21 accepts -std=c++26 cleanly and ships enough of the
draft library (std::print/std::println, std::expected, std::unreachable,
std::to_underlying, pack indexing) for us to move. The biggest
stylistic wins we'd have wanted (std::views::enumerate, ranges::zip)
aren't in this libc++ yet, so the refactor surface stays small.

Build configuration
  - CMakeLists.txt + tests/CMakeLists.txt: cxx_std_23 → cxx_std_26.
  - CMakeLists.txt: CMAKE_CXX_STANDARD 23 → 26.
  - tests/CMakeLists.txt: every e2e `-std=c++23` → `-std=c++26`.
  - .clang-format: Standard c++20 → Latest. (We had been parsing as
    C++20 even while building as C++23 — Latest tracks the compiler.)

Code refactors
  - sema/types: TypeArena::primitives_ is now keyed by `TypeKind`
    directly. std::hash works for any enum class since C++14, so the
    `static_cast<int>(k)` casts at insert and lookup are gone. This
    also tightens the key type to the enum's underlying width
    (uint16_t) instead of int, which is a wash size-wise but more
    honest.
  - tests/e2e/main_*.cpp: std::printf("compute() = %d\n", x) →
    std::println("compute() = {}", x). Same output, format-string
    type-checked at compile time, no more risk of a %d/%lld drift
    if a function's return type changes.

Documentation
  - README.md: every C++23 mention → C++26; toolchain matrix updated
    (Apple Clang 21 / LLVM Clang 19+ / GCC 15+); std::println /
    std::expected mentioned in the dependency list.
  - codegen header banner: C++23 → C++26.

Tests (109 passing, unchanged from before the bump; ran under both
Debug and AddressSanitizer+UBSan).

+++ src/sema/types.cpp
@@ -184,13 +184,13 @@ TypeArena::TypeArena() {
                    TypeKind::Bool,   TypeKind::Char,     TypeKind::Unit,    TypeKind::String,
                    TypeKind::Str,    TypeKind::StrConst, TypeKind::Never,   TypeKind::Error}) {
         auto t = std::unique_ptr<Type>(new Type(k));
-        primitives_.emplace(static_cast<int>(k), t.get());
+        primitives_.emplace(k, t.get());
         owned_.push_back(std::move(t));
     }
 }
 
 TypePtr TypeArena::primitive(TypeKind k) const {
-    auto it = primitives_.find(static_cast<int>(k));
+    auto it = primitives_.find(k);
     assert(it != primitives_.end() && "primitive(...) called with non-primitive kind");
     return it->second;
 }
